Adding all classes and a demo app on Main class
This is a small project saying about courses and students.
There is a course class and enrolling students in this course.
CourseRecord class includes student and his assignments and function
of calculating average and function to know if student can go to final
exam or not
Assignment class to set mask or get it
Student class contains his name